requirements.txt
django
djangorestframework
django-storages
boto3
python-dotenv

